[ska] relocate platform security & localization tests (#226465)
## Summary

Part of https://github.com/elastic/kibana-team/issues/1503

This PR is mostly about moving platform tests from x-pack/test

```
x-pack/test/
     | - functional_cloud/
     | - functional/apps/user_profiles
     | - localization/
```

After:
```
x-pack/platform/test/
     | - functional_cloud/
     | - functional/apps/user_profiles
     | - localization/
```

---------

(cherry picked from commit faae6ab00612327d3e49e142530c2bea6c9bf6dc)
