+++ .buildkite/ftr_platform_stateful_configs.yml
@@ -8,7 +8,7 @@ disabled:
   - src/platform/test/functional/firefox/config.base.ts
   - x-pack/test/functional/config.base.js
   - x-pack/platform/test/functional/config.base.ts
-  - x-pack/test/localization/config.base.ts
+  - x-pack/platform/test/localization/config.base.ts
   - src/platform/test/server_integration/config.base.js
   - x-pack/test/functional_with_es_ssl/config.base.ts
   - x-pack/test/api_integration/config.ts
@@ -134,9 +134,9 @@ enabled:
   - x-pack/test/accessibility/apps/group1/config.ts
   - x-pack/test/accessibility/apps/group2/config.ts
   - x-pack/test/accessibility/apps/group3/config.ts
-  - x-pack/test/localization/config.ja_jp.ts
-  - x-pack/test/localization/config.fr_fr.ts
-  - x-pack/test/localization/config.zh_cn.ts
+  - x-pack/platform/test/localization/config.ja_jp.ts
+  - x-pack/platform/test/localization/config.fr_fr.ts
+  - x-pack/platform/test/localization/config.zh_cn.ts
   - x-pack/platform/test/alerting_api_integration/basic/config.ts
   - x-pack/platform/test/alerting_api_integration/security_and_spaces/group1/config.ts
   - x-pack/platform/test/alerting_api_integration/security_and_spaces/group2/config.ts
@@ -263,7 +263,7 @@ enabled:
   - x-pack/test/functional/apps/transform/permissions/config.ts
   - x-pack/test/functional/apps/transform/feature_controls/config.ts
   - x-pack/platform/test/functional/apps/upgrade_assistant/config.ts
-  - x-pack/test/functional/apps/user_profiles/config.ts
+  - x-pack/platform/test/functional/apps/user_profiles/config.ts
   - x-pack/test/functional/apps/visualize/config.ts
   - x-pack/platform/test/functional/apps/watcher/config.ts
   - x-pack/platform/test/functional/config_security_basic.ts
@@ -271,8 +271,8 @@ enabled:
   - x-pack/test/functional/config.firefox.js
   - x-pack/platform/test/functional/config.firefox.ts
   - x-pack/platform/test/functional/config.upgrade_assistant.ts
-  - x-pack/test/functional_cloud/config.ts
-  - x-pack/test/functional_cloud/saml.config.ts
+  - x-pack/platform/test/functional_cloud/config.ts
+  - x-pack/platform/test/functional_cloud/saml.config.ts
   - x-pack/test/functional_solution_sidenav/config.ts
   - x-pack/test/licensing_plugin/config.public.ts
   - x-pack/test/licensing_plugin/config.ts
