+++ x-pack/platform/test/functional_cloud/config.ts
@@ -8,14 +8,17 @@ import { resolve } from 'path';
 import { FtrConfigProviderContext } from '@kbn/test';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
-  const functionalConfig = await readConfigFile(require.resolve('../functional/config.base.js'));
+  const functionalConfig = await readConfigFile(require.resolve('../functional/config.base.ts'));
 
   const kibanaPort = functionalConfig.get('servers.kibana.port');
   const idpPath = resolve(
     __dirname,
-    '../security_api_integration/plugins/saml_provider/metadata.xml'
+    '../../../test/security_api_integration/plugins/saml_provider/metadata.xml'
   );
-  const samlIdPPlugin = resolve(__dirname, '../security_api_integration/plugins/saml_provider');
+  const samlIdPPlugin = resolve(
+    __dirname,
+    '../../../test/security_api_integration/plugins/saml_provider'
+  );
 
   return {
     ...functionalConfig.getAll(),
