+++ x-pack/platform/test/functional_cloud/ftr_provider_context.ts
