+++ x-pack/platform/test/functional_cloud/saml.config.ts
@@ -26,11 +26,17 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const kibanaPort = kibanaFunctionalConfig.get('servers.kibana.port');
   const idpPath = resolve(
     __dirname,
-    '../security_api_integration/plugins/saml_provider/metadata.xml'
+    '../../../test/security_api_integration/plugins/saml_provider/metadata.xml'
   );
-  const samlIdPPlugin = resolve(__dirname, '../security_api_integration/plugins/saml_provider');
+  const samlIdPPlugin = resolve(
+    __dirname,
+    '../../../test/security_api_integration/plugins/saml_provider'
+  );
 
-  const testEndpointsPlugin = resolve(__dirname, '../security_functional/plugins/test_endpoints');
+  const testEndpointsPlugin = resolve(
+    __dirname,
+    '../../../test/security_functional/plugins/test_endpoints'
+  );
 
   return {
     testConfigCategory: ScoutTestRunConfigCategory.UI_TEST,
