+++ x-pack/platform/test/localization/config.base.ts
@@ -10,7 +10,7 @@ import { FtrConfigProviderContext } from '@kbn/test';
 import { services, pageObjects } from './ftr_provider_context';
 
 export async function withLocale({ readConfigFile }: FtrConfigProviderContext, locale: string) {
-  const functionalConfig = await readConfigFile(require.resolve('../functional/config.base.js'));
+  const functionalConfig = await readConfigFile(require.resolve('../functional/config.base.ts'));
 
   return {
     ...functionalConfig.getAll(),
