+++ x-pack/platform/test/localization/config.fr_fr.ts
