+++ x-pack/platform/test/localization/config.ja_jp.ts
