+++ x-pack/platform/test/localization/config.zh_cn.ts
