+++ x-pack/platform/test/localization/ftr_provider_context.ts
