+++ x-pack/platform/test/localization/tests/index.ts
