+++ x-pack/platform/test/localization/tests/lens/smokescreen.ts
