+++ x-pack/platform/test/localization/tests/login_page.ts
