+++ x-pack/test/functional/page_objects/index.ts
@@ -20,6 +20,7 @@ import { GisPageObject } from '@kbn/test-suites-xpack-platform/functional/page_o
 import { LensPageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/lens_page';
 import { UpgradeAssistantFlyoutObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/upgrade_assistant_page';
 import { SnapshotRestorePageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/snapshot_restore_page';
+import { UserProfilePageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/user_profile_page';
 import { ApiKeysPageProvider } from './api_keys_page';
 import { AssetDetailsProvider } from './asset_details';
 import { BannersPageObject } from './banners_page';
@@ -52,7 +53,6 @@ import { ShareSavedObjectsToSpacePageProvider } from './share_saved_objects_to_s
 import { StatusPageObject } from './status_page';
 import { TagManagementPageObject } from './tag_management_page';
 import { UptimePageObject } from './uptime_page';
-import { UserProfilePageProvider } from './user_profile_page';
 import { SearchPlaygroundPageProvider } from './search_playground_page';
 import { SearchSynonymsPageProvider } from './search_synonyms_page';
 import { SearchQueryRulesPageProvider } from './search_query_rules_page';
